+++ src/main/web/scoreEntry/scoreEntry.js
@@ -101,14 +101,16 @@ document.addEventListener("DOMContentLoaded", function() {
             document.getElementById("yesno-dialog").classList.remove("fll-sw-ui-inactive");
         });
     }
-    
-    
-        document.getElementById("score-entry_toggle-review-mode").addEventListener('click', () => {
+
+
+    document.getElementById("score-entry_toggle-review-mode").addEventListener('click', function() {
         const glassPane = document.getElementById("review-mode_glasspane");
         if (glassPane.style.zIndex > 0) {
             glassPane.style.zIndex = -1;
+            this.classList.remove("fll-sw-button-pressed");
         } else {
             glassPane.style.zIndex = 10;
+            this.classList.add("fll-sw-button-pressed");
         }
     });